WMHP-17: countdown hecho

+++ database/migrations/2023_09_14_125511_planta.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->enum('estado', ['Seca','Sana','Floreciendo','Brotando','Con manchas','Pocha','Muerta']);
             $table->string('nombre_planta');
             $table->string('cantidad_agua');
-            $table->decimal('riego');
+            $table->integer('riego');
         });
     }
